+++ src/main/java/com/E1i3/NoExit/domain/board/controller/BoardController.java
@@ -6,8 +6,8 @@ import com.E1i3.NoExit.domain.board.dto.BoardDetailResDto;
 import com.E1i3.NoExit.domain.board.dto.BoardListResDto;
 import com.E1i3.NoExit.domain.board.dto.BoardUpdateReqDto;
 import com.E1i3.NoExit.domain.board.service.BoardService;
-import com.E1i3.NoExit.domain.common.CommonErrorDto;
-import com.E1i3.NoExit.domain.common.CommonResDto;
+import com.E1i3.NoExit.domain.common.dto.CommonErrorDto;
+import com.E1i3.NoExit.domain.common.dto.CommonResDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
