+++ src/main/java/com/E1i3/NoExit/domain/comment/controller/CommentController.java
@@ -5,8 +5,8 @@ import com.E1i3.NoExit.domain.comment.dto.CommentCreateReqDto;
 import com.E1i3.NoExit.domain.comment.dto.CommentListResDto;
 import com.E1i3.NoExit.domain.comment.dto.CommentUpdateReqDto;
 import com.E1i3.NoExit.domain.comment.service.CommentService;
-import com.E1i3.NoExit.domain.common.CommonErrorDto;
-import com.E1i3.NoExit.domain.common.CommonResDto;
+import com.E1i3.NoExit.domain.common.dto.CommonErrorDto;
+import com.E1i3.NoExit.domain.common.dto.CommonResDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
